// frontend/src/types/generated/models/TriageDisposition.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Triage disposition for alerts
 */
export type TriageDisposition = 'TRUE_POSITIVE' | 'FALSE_POSITIVE' | 'BENIGN' | 'NEEDS_INVESTIGATION' | 'DUPLICATE' | 'UNKNOWN';